// src/tasks/proto/task.pb.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.7.7
//   protoc               v6.32.0
// source: task.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import { Observable } from "rxjs";

export const protobufPackage = "task";

export interface Task {
  id: number;
  title: string;
  description: string;
  completed: boolean;
  createdBy: number;
  createdAt: string;
}

export interface CreateTaskRequest {
  title: string;
  description: string;
  createdBy: string;
}

export interface GetTaskByIdRequest {
  id: number;
}

export interface UpdateTaskRequest {
  id: number;
  title: string;
  description: string;
  completed: boolean;
}

export interface DeleteTaskRequest {
  id: number;
}

export interface CompleteTaskRequest {
  id: number;
}

export interface Empty {
}

export interface GenericResponse {
  status: number;
  error?: string;
  task?: Task | undefined;
  taskList?: TaskList | undefined;
  deleteResult?: DeleteTaskResult | undefined;
}

export interface TaskList {
  tasks: Task[];
}

export interface DeleteTaskResult {
  success: boolean;
}

export const TASK_PACKAGE_NAME = "task";

export interface TaskServiceClient {
  createTask(request: CreateTaskRequest): Observable<GenericResponse>;

  getTaskById(request: GetTaskByIdRequest): Observable<GenericResponse>;

  getAllTasks(request: Empty): Observable<GenericResponse>;

  updateTask(request: UpdateTaskRequest): Observable<GenericResponse>;

  deleteTask(request: DeleteTaskRequest): Observable<GenericResponse>;

  completeTask(request: CompleteTaskRequest): Observable<GenericResponse>;
}

export interface TaskServiceController {
  createTask(request: CreateTaskRequest): Promise<GenericResponse> | Observable<GenericResponse> | GenericResponse;

  getTaskById(request: GetTaskByIdRequest): Promise<GenericResponse> | Observable<GenericResponse> | GenericResponse;

  getAllTasks(request: Empty): Promise<GenericResponse> | Observable<GenericResponse> | GenericResponse;

  updateTask(request: UpdateTaskRequest): Promise<GenericResponse> | Observable<GenericResponse> | GenericResponse;

  deleteTask(request: DeleteTaskRequest): Promise<GenericResponse> | Observable<GenericResponse> | GenericResponse;

  completeTask(request: CompleteTaskRequest): Promise<GenericResponse> | Observable<GenericResponse> | GenericResponse;
}

export function TaskServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = [
      "createTask",
      "getTaskById",
      "getAllTasks",
      "updateTask",
      "deleteTask",
      "completeTask",
    ];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("TaskService", method)(constructor.prototype[method], method, descriptor);
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("TaskService", method)(constructor.prototype[method], method, descriptor);
    }
  };
}

export const TASK_SERVICE_NAME = "TaskService";